# specs/001-personality-test/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: MBTI Personality Test API
  description: API for the 16 personality types test based on MBTI methodology
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3001/api
    description: Development server
paths:
...
  /questions/current:
    get:
      summary: Get current question
      description: Returns the current question based on session progress
      operationId: getCurrentQuestion
      tags:
        - Questions
      responses:
        '200':
          description: Current question with progress info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
        '400':
          description: Test already completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No active session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /questions/{questionId}:
    get:
      summary: Get specific question
      description: Returns a specific question by ID (for navigation back)
      operationId: getQuestion
      tags:
        - Questions
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 20
      responses:
        '200':
          description: Question with answer if previously answered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
        '400':
          description: Question not yet accessible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Question not found or no active session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    QuestionResponse:
      type: object
      required:
        - question
        - currentIndex
        - totalQuestions
      properties:
        question:
          $ref: '#/components/schemas/Question'
        currentIndex:
          type: integer
          minimum: 0
          maximum: 19
          description: Current question index (0-based)
        totalQuestions:
          type: integer
          example: 20
        previousAnswer:
          $ref: '#/components/schemas/Pole'
          nullable: true
          description: Previously selected answer for this question, if any
    Question:
      type: object
      required:
        - id
        - text
        - options
      properties:
        id:
          type: integer
          minimum: 1
          maximum: 20
        text:
          type: string
          maxLength: 500
        options:
          type: array
          items:
            $ref: '#/components/schemas/QuestionOption'
          minItems: 2
          maxItems: 2
    QuestionOption:
      type: object
      required:
        - text
        - pole
      properties:
        text:
          type: string
        pole:
          $ref: '#/components/schemas/Pole'
    Pole:
      type: string
      enum:
        - E
        - I
        - S
        - N
        - T
        - F
        - J
        - P
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
security:
  - sessionCookie: []
...